requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
# sqlalchemy==1.4.25
# psycopg2
# mysqlclient
pymysql
python-dotenv
cryptography